add binding User from DataBase, and update labels whit data

+++ Models/TwitterUser.swift
@@ -12,8 +12,8 @@ struct TwitterUser: Codable {
     let id: String
     var displayname: String = ""
     var userName: String = ""
-    var followingCount: Double = 0
-    var followersCount: Double = 0
+    var followingCount: Int = 0
+    var followersCount: Int = 0
     var createdOn: Date = Date()
     var bio: String = ""
     var avatarPath: String = ""
